+++ src/main/java/com/robertx22/age_of_exile/aoe_data/database/base_gear_types/adders/BaseClothArmors.java
@@ -13,8 +13,12 @@ import com.robertx22.age_of_exile.database.registry.ISlashRegistryInit;
 import com.robertx22.age_of_exile.mmorpg.ModRegistry;
 import com.robertx22.age_of_exile.saveclasses.gearitem.gear_bases.StatRequirement;
 import com.robertx22.age_of_exile.uncommon.enumclasses.ModType;
+import net.minecraft.entity.EquipmentSlot;
 
 import java.util.HashMap;
+
+import static com.robertx22.age_of_exile.uncommon.utilityclasses.SlotUtils.multiOf;
+import static net.minecraft.entity.EquipmentSlot.FEET;
 
 public class BaseClothArmors implements ISlashRegistryInit {
 
@@ -23,34 +27,37 @@ public class BaseClothArmors implements ISlashRegistryInit {
     public static HashMap<LevelRange, DataGenKey<BaseGearType>> CHESTS;
     public static HashMap<LevelRange, DataGenKey<BaseGearType>> HELMETS;
 
+    static float min = 4;
+    static float max = 8;
+
     @Override
     public void registerAll() {
 
         BOOTS = BaseGearBuilder.of(GearSlots.BOOTS, "cloth_boots", "Shoes", ModRegistry.GEAR_ITEMS.CLOTH_BOOTS)
             .req(new StatRequirement().intelligence(0.5f))
             .tags(new TagList(SlotTag.cloth, SlotTag.boots, SlotTag.armor_family, SlotTag.magic_shield_stat, SlotTag.intelligence))
-            .baseStat(new StatModifier(1, 3, MagicShield.getInstance(), ModType.FLAT))
+            .baseStat(new StatModifier(multiOf(FEET) * min, multiOf(FEET) * max, MagicShield.getInstance(), ModType.FLAT))
             .addMageLevelRanges()
             .build();
 
         PANTS = BaseGearBuilder.of(GearSlots.PANTS, "cloth_pants", "Leggings", ModRegistry.GEAR_ITEMS.CLOTH_PANTS)
             .req(new StatRequirement().intelligence(0.5f))
             .tags(new TagList(SlotTag.cloth, SlotTag.pants, SlotTag.armor_family, SlotTag.magic_shield_stat, SlotTag.intelligence))
-            .baseStat(new StatModifier(3, 7, MagicShield.getInstance(), ModType.FLAT))
+            .baseStat(new StatModifier(multiOf(EquipmentSlot.LEGS) * min, multiOf(EquipmentSlot.LEGS) * max, MagicShield.getInstance(), ModType.FLAT))
             .addMageLevelRanges()
             .build();
 
         CHESTS = BaseGearBuilder.of(GearSlots.CHEST, "cloth_chest", "Robe", ModRegistry.GEAR_ITEMS.CLOTH_CHESTS)
             .req(new StatRequirement().intelligence(0.5f))
             .tags(new TagList(SlotTag.cloth, SlotTag.chest, SlotTag.armor_family, SlotTag.magic_shield_stat, SlotTag.intelligence))
-            .baseStat(new StatModifier(4, 8, MagicShield.getInstance(), ModType.FLAT))
+            .baseStat(new StatModifier(multiOf(EquipmentSlot.CHEST) * min, multiOf(EquipmentSlot.CHEST) * max, MagicShield.getInstance(), ModType.FLAT))
             .addMageLevelRanges()
             .build();
 
         HELMETS = BaseGearBuilder.of(GearSlots.HELMET, "cloth_helmet", "Hat", ModRegistry.GEAR_ITEMS.CLOTH_HELMETS)
             .req(new StatRequirement().intelligence(0.5f))
             .tags(new TagList(SlotTag.cloth, SlotTag.helmet, SlotTag.armor_family, SlotTag.magic_shield_stat, SlotTag.intelligence))
-            .baseStat(new StatModifier(2, 5, MagicShield.getInstance(), ModType.FLAT))
+            .baseStat(new StatModifier(multiOf(EquipmentSlot.HEAD) * min, multiOf(EquipmentSlot.HEAD) * max, MagicShield.getInstance(), ModType.FLAT))
             .addMageLevelRanges()
             .build();
     }